// src/utils/enums/email/templates.enum.ts
//set title as search term to get ids from env
export enum MailTemplateIds {
  SEND_OTP = 'YOUR_OTP_CODE',
  FORGOT_PASSWORD = 'FORGOT_PASSWORD',
  REGISTER_PASSWORD = 'REGISTER_PASSWORD',
  SEND_TEXT = 'SEND_TEXT',
  LEVERAGE_CHANGE = 'CHANGE_LEVERAGE',
  MT5_ACCOUNT = 'MT5_ACCOUNT_CREATION',
  TRADING_ACCOUNT_PASSWORD_LIVE = 'TRADING_ACCOUNT_PASSWORD_LIVE',
  TRADING_ACCOUNT_PASSWORD_DEMO = 'TRADING_ACCOUNT_PASSWORD_DEMO',
  SEND_LEAD_ASSIGNMENT_EMAIL = 'SEND_LEAD_ASSIGNMENT_EMAIL',
  SEND_CLIENT_ASSIGNMENT_EMAIL = 'SEND_CLIENT_ASSIGNMENT_EMAIL',
}